+++ src/main/java/com/example/jsgamesbackendmain/Bean/GameBean/GameCreateBean.java
@@ -1,24 +1,20 @@
 package com.example.jsgamesbackendmain.Bean.GameBean;
 
 import com.example.jsgamesbackendmain.Model.DAO.GameDAO;
+import com.example.jsgamesbackendmain.Model.DTO.Game.GameCreateRequestDTO;
 import com.example.jsgamesbackendmain.Model.DTO.Game.GameDTO;
-import com.example.jsgamesbackendmain.Model.DTO.Game.GameListResponseDTO;
 import com.example.jsgamesbackendmain.Repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Component
-public class Game2Bean {
+public class GameCreateBean {
     @Autowired
     private GameRepository gameRepository;
 
-    public GameListResponseDTO exec() {
-        List<GameDAO> gameList = gameRepository.findAll();
-        List<GameDTO> dtoList = gameList.stream().map(GameDTO::of).collect(Collectors.toList());
+    public GameDTO exec(GameCreateRequestDTO gameCreateRequestDTO) {
+        GameDAO savedGame = gameRepository.save(GameCreateRequestDTO.toGameDAO(gameCreateRequestDTO));
 
-        return GameListResponseDTO.of(dtoList, "SUCCESS");
+        return GameDTO.of(savedGame);
     }
 }
