+++ src/main/java/com/example/jsgamesbackendmain/Service/GameService.java
@@ -1,8 +1,8 @@
 package com.example.jsgamesbackendmain.Service;
 
-import com.example.jsgamesbackendmain.Bean.GameBean.Game1Bean;
-import com.example.jsgamesbackendmain.Bean.GameBean.Game2Bean;
-import com.example.jsgamesbackendmain.Bean.GameBean.Game3Bean;
+import com.example.jsgamesbackendmain.Bean.GameBean.GameCreateBean;
+import com.example.jsgamesbackendmain.Bean.GameBean.GameListBean;
+import com.example.jsgamesbackendmain.Bean.GameBean.GameGetBean;
 import com.example.jsgamesbackendmain.Model.DTO.Game.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,15 +10,15 @@ import org.springframework.stereotype.Service;
 @Service
 public class GameService {
     @Autowired
-    private Game1Bean game1Bean;
+    private GameCreateBean game1Bean;
     @Autowired
-    private Game2Bean game2Bean;
+    private GameListBean game2Bean;
     @Autowired
-    private Game3Bean game3Bean;
+    private GameGetBean game3Bean;
 
-    public GameCreateResponseDTO addGame(GameCreateRequestDTO gameCreateRequestDTO) { return game1Bean.exec(gameCreateRequestDTO); }
+    public GameDTO createGame(GameCreateRequestDTO gameCreateRequestDTO) { return game1Bean.exec(gameCreateRequestDTO); }
 
     public GameListResponseDTO listGames() { return game2Bean.exec(); }
 
-    public GameDetailResponseDTO getGame(Long gameId) { return game3Bean.exec(gameId); }
+    public GameDTO getGame(Long gameId) { return game3Bean.exec(gameId); }
 }
